+++ kernel/src/ipc/mod.rs
@@ -3,4 +3,5 @@ pub mod pipe;
 pub mod shm;
 pub mod signal;
 pub mod signal_types;
+pub mod kill;
 pub mod syscall;
